@@ -23,7 +23,10 @@ import 'package:joyful_noise/core/presentation/bootstrap.dart';
 
 class SongDetailPage extends ConsumerStatefulWidget {
   final Song song;
-  const SongDetailPage({Key? key, required this.song}) : super(key: key);
+  const SongDetailPage({
+    required this.song,
+    super.key,
+  });
 
   @override
   SongDetailPageState createState() => SongDetailPageState();
@@ -132,7 +135,7 @@ class SongDetailPageState extends ConsumerState<SongDetailPage> {
                         ref.read(songDetailNotifierProvider.notifier).switchStarredStatus(state.songDetail.entity!);
                         ref.refresh(favoriteSongsNotifierProvider.notifier).getFirstFavoriteSongsPage();
                       },
-                icon: state.songDetail.entity?.isFavorite == true
+                icon: state.songDetail.entity?.isFavorite ?? true
                     ? const Icon(Icons.star)
                     : const Icon(Icons.star_outline),
               );
@@ -142,7 +145,7 @@ class SongDetailPageState extends ConsumerState<SongDetailPage> {
       ),
       body: InteractiveViewer(
         panEnabled: false,
-        minScale: 0.5,
+        minScale: 0.01,
         boundaryMargin: const EdgeInsets.all(80),
         maxScale: 4,
         child: Column(
