@@ -1,48 +1,48 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:joyful_noise/backend/songs/core/presentation/paginated_songs_list_view.dart';
+import 'package:joyful_noise/main_development.dart' as main_dart;
 import 'package:flutter_test/flutter_test.dart';
-import 'package:integration_test/integration_test.dart';
-
-import 'package:joyful_noise/auth/presentation/authorization_page.dart';
-import 'package:mocktail/mocktail.dart';
-import 'package:webview_flutter/webview_flutter.dart';
-
-import '../test/auth/presentation/authorization_page_test.dart';
-
-class MockNavigatorObserver extends Mock implements NavigatorObserver {}
+import 'package:joyful_noise/search/presentation/search_bar.dart';
+import 'package:patrol/patrol.dart';
 
 void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+  patrolTest(
+    'counter state is the same after going to home and switching apps',
+    nativeAutomation: true,
+    ($) async {
+      final originalOnError = FlutterError.onError;
+      main_dart.main();
+      for (var i = 0; i < 300; i++) {
+        await $.pump();
+      }
+      FlutterError.onError = originalOnError;
 
-  group('WebViewExample', () {
-    testWidgets('fill the email and password text fields', (WidgetTester tester) async {
-      final mockObserver = MockNavigatorObserver();
-
-      final OnAuthorizationCodeRedirectAttemptCallback mockOnAuthorizationCodeRedirectAttemptCallback =
-          MockOnAuthorizationCodeRedirectAttemptCallback();
-
-      when(mockOnAuthorizationCodeRedirectAttemptCallback.call).thenReturn((_) {});
-
-      await tester.pumpWidget(
-        MaterialApp(
-          home: AuthorizationPage(
-            authorizationUrl: Uri.parse(
-              'https://mdbootstrap.com/docs/standard/extended/login/#section-basic-example',
-            ),
-            onAuthorizationCodeRedirectAttempt: mockOnAuthorizationCodeRedirectAttemptCallback(),
-            //   onWebViewCreatedJsString: """
-            // document.getElementById('form2Example1').value='example.com'
-            // document.getElementsByClassName("form-label")[0].innerHTML = "";
-            // document.getElementById('form2Example2').value='password'
-            // document.getElementsByClassName("form-label")[1].innerHTML = "";
-            // document.getElementsByClassName("btn btn-primary btn-block mb-4")[0].click();
-            // """,
-          ),
-          navigatorObservers: [mockObserver],
-        ),
+      await $(#signInButtonKey).tap();
+      for (var i = 0; i < 300; i++) {
+        await $.pump();
+      }
+      await $.native.enterTextByIndex(
+        'hey@hey.com',
+        index: 0,
       );
-      await tester.pumpAndSettle();
+      await $.native.enterTextByIndex(
+        'heyheyhey',
+        index: 1,
+      );
 
-      expect(find.byType(WebViewWidget), findsOneWidget);
-    });
-  });
+      await $.native.tap(Selector(text: 'Sign in'));
+
+      for (var i = 0; i < 300; i++) {
+        await $.pump();
+      }
+
+      expect($(SearchBar), findsOneWidget);
+      expect($(PaginatedSongsListView), findsOneWidget);
+      await $(#signOutButtonKey).tap();
+      for (var i = 0; i < 300; i++) {
+        await $.pump();
+      }
+    },
+  );
 }
