@@ -9,10 +9,10 @@ import 'package:joyful_noise/auth/infrastructure/webapp_authenticator.dart';
 
 class AuthorizationPage extends StatefulWidget {
   const AuthorizationPage({
-    Key? key,
     required this.authorizationUrl,
     required this.onAuthorizationCodeRedirectAttempt,
-  }) : super(key: key);
+    super.key,
+  });
 
   static const backButtonKey = ValueKey('backButton');
 
