@@ -1,9 +1,7 @@
 // Package imports:
-import 'package:dio/dio.dart';
 
 // Project imports:
 import 'package:joyful_noise/backend/core/infrastructure/backend_base_url.dart';
-import 'package:joyful_noise/backend/core/infrastructure/backend_headers_cache.dart';
 import 'package:joyful_noise/backend/core/infrastructure/pagination_config.dart';
 import 'package:joyful_noise/backend/core/infrastructure/song_dto.dart';
 import 'package:joyful_noise/backend/songs/core/infrastructure/songs_page_remote_service.dart';
@@ -11,9 +9,9 @@ import 'package:joyful_noise/core/infrastructure/remote_response.dart';
 
 class SearchedSongsRemoteService extends SongsPageRemoteService {
   SearchedSongsRemoteService(
-    Dio dio,
-    BackendHeadersCache headersCache,
-  ) : super(dio, headersCache);
+    super.dio,
+    super.headersCache,
+  );
 
   Future<RemoteResponse<List<SongDTO>>> getSearchedSongsPage(
     String query,
@@ -21,7 +19,6 @@ class SearchedSongsRemoteService extends SongsPageRemoteService {
   ) async =>
       super.getPage(
         storeEtag: false,
-        // TODO(jeremiah): http for local https for real
         requestUri: Uri.https(
           BackendConstants().backendBaseUrl(),
           '/api/v1/songs',
