@@ -6,6 +6,6 @@ import 'package:dio/dio.dart';
 
 extension DioErrorX on DioError {
   bool get isNoConnectionError {
-    return type == DioErrorType.other && error is SocketException;
+    return type == DioErrorType.unknown && error is SocketException;
   }
 }
