@@ -28,13 +28,13 @@ class SearchBar extends ConsumerStatefulWidget {
   final void Function() onSignOutButtonPressed;
 
   const SearchBar({
-    Key? key,
     required this.body,
     required this.title,
     required this.hint,
     required this.onShouldNavigateToResultPage,
     required this.onSignOutButtonPressed,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   SearchBarState createState() => SearchBarState();
