@@ -35,7 +35,7 @@ void main() {
         ).thenAnswer(
           (invocation) => Future.value(
             Response<dynamic>(
-              requestOptions: RequestOptions(path: ''),
+              requestOptions: RequestOptions(),
               statusCode: 304,
             ),
           ),
@@ -66,7 +66,7 @@ void main() {
         ).thenAnswer(
           (invocation) => Future.value(
             Response<dynamic>(
-              requestOptions: RequestOptions(path: ''),
+              requestOptions: RequestOptions(),
               statusCode: 200,
               data: mockData,
             ),
@@ -98,7 +98,7 @@ void main() {
         ).thenAnswer(
           (invocation) => Future.value(
             Response<dynamic>(
-              requestOptions: RequestOptions(path: ''),
+              requestOptions: RequestOptions(),
               statusCode: 400,
             ),
           ),
@@ -124,7 +124,7 @@ void main() {
           () => mockDio.getUri<dynamic>(any(), options: any(named: 'options')),
         ).thenThrow(
           DioError(
-            requestOptions: RequestOptions(path: ''),
+            requestOptions: RequestOptions(),
             error: const SocketException(''),
           ),
         );
@@ -149,8 +149,8 @@ void main() {
           () => mockDio.getUri<dynamic>(any(), options: any(named: 'options')),
         ).thenThrow(
           DioError(
-            requestOptions: RequestOptions(path: ''),
-            response: Response<dynamic>(requestOptions: RequestOptions(path: '')),
+            requestOptions: RequestOptions(),
+            response: Response<dynamic>(requestOptions: RequestOptions()),
           ),
         );
 
